Attack Animation
Added the attack animation to the game. Cleaned up some off the code and
added comments for the logic.

There is a bug that if the idle animation doesn't start after attacking,
the attack animation continues where it left off. Going to work on
fixing that at a later date.